feature-policy/container-policy-fullscreen-manual.sub.html

<!DOCTYPE html>
<title>Fullscreen allowed by container policy</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="f1" allow="fullscreen"></iframe>
<iframe id="f2" allow="fullscreen" allowfullscreen></iframe>
<script>
var srcs = [
  'resources/fullscreen.html',
  '{{location[scheme]}}://{{domains[www]}}:{{location[port]}}/feature-policy/resources/fullscreen.html'
];

function loadFrame(id) {
  var frame = document.getElementById(id);
  for (var src of srcs) {
    promise_test(function(t) {
      frame.src = src;
      return new Promise(function(resolve, reject) {
        window.addEventListener('message', function(e) {
          resolve(e.data);
        }, { once: true });
      }).then(function(data) {
        assert_true(data.enabled, 'Document.fullscreenEnabled:');
        assert_equals(data.type, 'change', 'Document.requestFullscreen():');
      });
  }, (src === srcs[0] ? 'same origin' : 'cross origin') +
    ' iframe with allowfullscreen = ' + frame.allowFullscreen +
    ' may call document.requestFullscreen() when enabled by feature policy');
  }
}

loadFrame('f1');
loadFrame('f2');
</script>
